continue writing tests, fixing bugs
* made a number of tests depend simply on LeafValue
* fixed right-rotation pointer bugs
* fixed implementation of rank for u128 with indices above 64
* attempted implementation of `StaticBitVec` for Node (failed)
* adaptive dotviz for all current LeafValue sizes
* added ascii-art for rotation functions

+++ src/leaf/mod.rs
@@ -8,7 +8,8 @@ type Side<T> = either::Either<T, T>;
 /// Container type used to contain bits in [`Leaf`]. Sensible options are [`u64`] and [`u128`].
 /// Might be replaced with custom implementation featuring higher bit container size later (e.g.
 /// 4096, or dynamically dependent on total BitVector capacity).
-pub type LeafValue = u64;
+// pub type LeafValue = u64;
+pub type LeafValue = u128;
 
 pub const HALF: u32 = LeafValue::BITS / 2;
